feat(search-bar): added shimmer
